refactor(tui): split app/mod.rs monolith into focused submodules
Split crates/zeph-tui/src/app/mod.rs (4235 lines) into focused submodules,
joining the existing draw/events/keys split:

- state.rs: App construction, builder configuration, and accessors
- transcript.rs: sub-agent transcript loading, polling, and projection
- tests.rs: the 188-test #[cfg(test)] module (~67% of the file)

mod.rs (660 lines) retains the type definitions (App, Panel, AgentViewTarget,
TuiTranscriptEntry, TranscriptCache, SubAgentSidebarState, ConfirmState,
ElicitationState) and the shared free functions. The shared impl App block is
now split across files, so three cross-module methods (trim_messages,
auto_scroll, maybe_reload_transcript) are pub(super). Production code is
byte-identical; no public API change.

Closes #4918
