+++ crates/zeph-tui/src/app/state.rs
@@ -1,0 +1,610 @@
+// SPDX-FileCopyrightText: 2026 Andrei G <bug-ops>
+// SPDX-License-Identifier: MIT OR Apache-2.0
+
+//! App construction, builder configuration, and accessors over the active session
+//! state (input, messages, scroll, panels, metrics, and display toggles).
+
+use std::sync::Arc;
+
+use tokio::sync::{Notify, mpsc, watch};
+use zeph_common::task_supervisor::TaskSupervisor;
+
+use crate::command::TuiCommand;
+use crate::event::AgentEvent;
+use crate::hyperlink::HyperlinkSpan;
+use crate::metrics::MetricsSnapshot;
+use crate::session::SessionRegistry;
+use crate::types::PasteState;
+use crate::widgets::tool_view::ToolDensity;
+
+use super::{
+    AgentViewTarget, App, ChatMessage, InputMode, MAX_VISIBLE_INPUT_LINES, MessageRole, Panel,
+    RenderCache, SubAgentSidebarState, TranscriptCache, is_tool_use_only, parse_tool_output,
+};
+
+impl App {
+    /// Create a new `App` with the given I/O channels.
+    ///
+    /// The app starts in insert mode with the splash screen visible and no
+    /// messages in the buffer.
+    ///
+    /// # Arguments
+    ///
+    /// * `user_input_tx` — sender used to forward the user's typed text to the
+    ///   agent loop via [`TuiChannel`](crate::TuiChannel).
+    /// * `agent_event_rx` — receiver for [`AgentEvent`] produced by the agent.
+    ///
+    /// # Examples
+    ///
+    /// ```rust
+    /// use tokio::sync::mpsc;
+    /// use zeph_tui::App;
+    ///
+    /// let (user_tx, _user_rx) = mpsc::channel(64);
+    /// let (_agent_tx, agent_rx) = mpsc::channel(64);
+    /// let app = App::new(user_tx, agent_rx);
+    /// assert!(app.show_splash());
+    /// ```
+    #[must_use]
+    pub fn new(
+        user_input_tx: mpsc::Sender<String>,
+        agent_event_rx: mpsc::Receiver<AgentEvent>,
+    ) -> Self {
+        Self {
+            sessions: SessionRegistry::bootstrap(),
+            show_side_panels: true,
+            show_help: false,
+            metrics: MetricsSnapshot::default(),
+            metrics_rx: None,
+            active_panel: Panel::Chat,
+            tool_expanded: false,
+            tool_density: ToolDensity::default(),
+            show_source_labels: false,
+            show_balance: true,
+            throbber_state: throbber_widgets_tui::ThrobberState::default(),
+            confirm_state: None,
+            elicitation_state: None,
+            command_palette: None,
+            command_tx: None,
+            file_picker_state: None,
+            file_index: None,
+            slash_autocomplete: None,
+            reverse_search: None,
+            should_quit: false,
+            user_input_tx,
+            agent_event_rx,
+            queued_count: 0,
+            pending_count: 0,
+            context_token_estimate: 0,
+            editing_queued: false,
+            hyperlinks: Vec::new(),
+            cancel_signal: None,
+            pending_file_index: None,
+            subagent_sidebar: SubAgentSidebarState::new(),
+            task_supervisor: None,
+            show_task_panel: false,
+            cached_task_snapshots: Vec::new(),
+            clipboard: crate::clipboard::ClipboardHandle::new(),
+            fleet_snapshot: crate::widgets::fleet::FleetSnapshot::default(),
+            fleet_list_state: ratatui::widgets::ListState::default(),
+        }
+    }
+
+    /// Return `true` while the splash screen should be displayed.
+    ///
+    /// The splash screen is hidden as soon as the first chat message arrives.
+    #[must_use]
+    pub fn show_splash(&self) -> bool {
+        self.sessions.current().show_splash
+    }
+
+    /// Return `true` when the side panels column is visible.
+    ///
+    /// Controlled by the `s` keybinding and automatically disabled on narrow
+    /// terminals (< 80 columns).
+    #[must_use]
+    pub fn show_side_panels(&self) -> bool {
+        self.show_side_panels
+    }
+
+    /// Returns `true` when the user has toggled back to subagents view (plan view overridden).
+    #[must_use]
+    pub fn plan_view_active(&self) -> bool {
+        self.sessions.current().plan_view_active
+    }
+
+    // ---- Accessors for fields relocated into SessionSlot (preserves pub API surface) ----
+
+    /// Returns the active session's render cache.
+    #[must_use]
+    pub fn render_cache(&self) -> &RenderCache {
+        &self.sessions.current().render_cache
+    }
+
+    /// Returns a mutable reference to the active session's render cache.
+    pub fn render_cache_mut(&mut self) -> &mut RenderCache {
+        &mut self.sessions.current_mut().render_cache
+    }
+
+    /// Returns the current chat area view target (main conversation or sub-agent transcript).
+    #[must_use]
+    pub fn view_target(&self) -> &AgentViewTarget {
+        &self.sessions.current().view_target
+    }
+
+    /// Returns the cached transcript for the currently-focused sub-agent, if any.
+    #[must_use]
+    pub fn transcript_cache(&self) -> Option<&TranscriptCache> {
+        self.sessions.current().transcript_cache.as_ref()
+    }
+
+    /// Populate the message buffer from a persisted session history.
+    ///
+    /// Each element is a `(role, content)` pair where `role` is one of
+    /// `"user"`, `"assistant"`, or `"tool"`. Tool outputs are detected by a
+    /// sentinel suffix and rendered as [`MessageRole::Tool`] messages.
+    /// The splash screen is hidden after loading if any messages are present.
+    pub fn load_history(&mut self, messages: &[(&str, &str)]) {
+        const TOOL_SUFFIX: &str = "\n```";
+
+        for &(role_str, content) in messages {
+            if role_str == "user"
+                && let Some((tool_name, body)) = parse_tool_output(content, TOOL_SUFFIX)
+            {
+                self.sessions
+                    .current_mut()
+                    .messages
+                    .push(ChatMessage::new(MessageRole::Tool, body).with_tool(tool_name.into()));
+                continue;
+            }
+
+            let role = match role_str {
+                "user" => MessageRole::User,
+                "assistant" => {
+                    if is_tool_use_only(content) {
+                        continue;
+                    }
+                    MessageRole::Assistant
+                }
+                _ => continue,
+            };
+            if role == MessageRole::User {
+                self.sessions
+                    .current_mut()
+                    .input_history
+                    .push(content.to_owned());
+            }
+            self.sessions
+                .current_mut()
+                .messages
+                .push(ChatMessage::new(role, content));
+        }
+        // Enforce the message buffer cap on initial history load as well.
+        self.trim_messages();
+        if !self.sessions.current().messages.is_empty() {
+            self.sessions.current_mut().show_splash = false;
+        }
+    }
+
+    /// Attach a cancel signal that Ctrl-C in the TUI will trigger.
+    ///
+    /// # Examples
+    ///
+    /// ```rust
+    /// use std::sync::Arc;
+    /// use tokio::sync::{Notify, mpsc};
+    /// use zeph_tui::App;
+    ///
+    /// let (tx, _rx) = mpsc::channel(1);
+    /// let (_atx, arx) = mpsc::channel(1);
+    /// let notify = Arc::new(Notify::new());
+    /// let _app = App::new(tx, arx).with_cancel_signal(notify);
+    /// ```
+    #[must_use]
+    pub fn with_cancel_signal(mut self, signal: Arc<Notify>) -> Self {
+        self.cancel_signal = Some(signal);
+        self
+    }
+
+    /// Attach a metrics watch channel for live dashboard updates.
+    ///
+    /// The current snapshot is read immediately; subsequent updates are polled
+    /// by [`poll_metrics`](Self::poll_metrics) each frame.
+    ///
+    /// # Examples
+    ///
+    /// ```rust
+    /// use tokio::sync::{mpsc, watch};
+    /// use zeph_tui::{App, MetricsSnapshot};
+    ///
+    /// let (tx, _rx) = mpsc::channel(1);
+    /// let (_atx, arx) = mpsc::channel(1);
+    /// let (_metrics_tx, metrics_rx) = watch::channel(MetricsSnapshot::default());
+    /// let _app = App::new(tx, arx).with_metrics_rx(metrics_rx);
+    /// ```
+    #[must_use]
+    pub fn with_metrics_rx(mut self, rx: watch::Receiver<MetricsSnapshot>) -> Self {
+        self.metrics = rx.borrow().clone();
+        self.metrics_rx = Some(rx);
+        self
+    }
+
+    /// Attach the command dispatch sender used for slash-command routing.
+    ///
+    /// # Examples
+    ///
+    /// ```rust
+    /// use tokio::sync::mpsc;
+    /// use zeph_tui::{App, TuiCommand};
+    ///
+    /// let (tx, _rx) = mpsc::channel(1);
+    /// let (_atx, arx) = mpsc::channel(1);
+    /// let (cmd_tx, _cmd_rx) = mpsc::channel(8);
+    /// let _app = App::new(tx, arx).with_command_tx(cmd_tx);
+    /// ```
+    #[must_use]
+    pub fn with_command_tx(mut self, tx: mpsc::Sender<TuiCommand>) -> Self {
+        self.command_tx = Some(tx);
+        self
+    }
+
+    /// Set the initial tool-output density from a loaded `TuiConfig`.
+    ///
+    /// Applied once at startup; runtime changes via the `c` key override this
+    /// but are not persisted back to config.
+    ///
+    /// # Examples
+    ///
+    /// ```rust
+    /// use tokio::sync::mpsc;
+    /// use zeph_tui::App;
+    /// use zeph_config::ToolDensity;
+    ///
+    /// let (tx, _rx) = mpsc::channel(1);
+    /// let (_atx, arx) = mpsc::channel(1);
+    /// let _app = App::new(tx, arx).with_tool_density(ToolDensity::Compact);
+    /// ```
+    #[must_use]
+    pub fn with_tool_density(mut self, density: ToolDensity) -> Self {
+        self.tool_density = density;
+        self
+    }
+
+    /// Wire a [`TaskSupervisor`] into the `App` for the task registry panel.
+    ///
+    /// The supervisor's task list is snapshotted once per render tick before
+    /// `terminal.draw()`, keeping the draw closure free of mutex contention.
+    /// Toggle the panel visibility with `/tasks`.
+    ///
+    /// # Examples
+    ///
+    /// ```rust,ignore
+    /// use tokio::sync::mpsc;
+    /// use tokio_util::sync::CancellationToken;
+    /// use zeph_common::task_supervisor::TaskSupervisor;
+    /// use zeph_tui::App;
+    ///
+    /// let (user_tx, _) = mpsc::channel(64);
+    /// let (_, agent_rx) = mpsc::channel(64);
+    /// let cancel = CancellationToken::new();
+    /// let supervisor = TaskSupervisor::new(cancel);
+    /// let _app = App::new(user_tx, agent_rx).with_task_supervisor(supervisor);
+    /// ```
+    #[must_use]
+    pub fn with_task_supervisor(mut self, supervisor: TaskSupervisor) -> Self {
+        self.task_supervisor = Some(supervisor);
+        self
+    }
+
+    /// Refresh the cached task snapshot from the supervisor.
+    ///
+    /// Must be called once per render tick **before** `terminal.draw()` to avoid
+    /// acquiring the supervisor's inner mutex inside the draw closure.
+    pub(crate) fn refresh_task_snapshots(&mut self) {
+        self.cached_task_snapshots = self
+            .task_supervisor
+            .as_ref()
+            .map(TaskSupervisor::snapshot)
+            .unwrap_or_default();
+    }
+
+    /// Return a truncated label for active `TaskSupervisor` tasks, or `None` when idle.
+    ///
+    /// Used by the input widget to show a braille spinner with the name of the first
+    /// active (Running/Restarting) task when no other status is being displayed.
+    #[must_use]
+    pub fn supervisor_activity_label(&self) -> Option<String> {
+        self.task_supervisor.as_ref()?;
+        let mut active = self
+            .cached_task_snapshots
+            .iter()
+            .filter(|t| {
+                matches!(
+                    t.status,
+                    zeph_common::task_supervisor::TaskStatus::Running
+                        | zeph_common::task_supervisor::TaskStatus::Restarting { .. }
+                )
+            })
+            .filter(|t| !t.name.starts_with("mem-"))
+            .peekable();
+        let first = active.next()?;
+        let label = if active.peek().is_none() {
+            first.name.to_string()
+        } else {
+            let extra = active.count() + 1; // +1 because we already consumed first
+            format!("{} +{} more", first.name, extra)
+        };
+        // Char-based truncation to avoid panicking on multi-byte UTF-8 boundaries.
+        let truncated: String = label.chars().take(38).collect();
+        Some(truncated)
+    }
+
+    /// Wire a cancel signal into a running App instance.
+    ///
+    /// Used by the two-phase TUI startup path to connect the agent's cancel signal
+    /// after the agent has been constructed (Phase 2).
+    pub fn set_cancel_signal(&mut self, signal: Arc<Notify>) {
+        self.cancel_signal = Some(signal);
+    }
+
+    /// Wire a metrics receiver into a running App instance.
+    ///
+    /// Used by the two-phase TUI startup path to connect the metrics channel
+    /// after the metrics watch channel has been created (Phase 2).
+    pub fn set_metrics_rx(&mut self, rx: watch::Receiver<MetricsSnapshot>) {
+        self.metrics = rx.borrow().clone();
+        self.metrics_rx = Some(rx);
+    }
+
+    /// Check the metrics watch channel for an updated snapshot and apply it.
+    ///
+    /// Also clamps the sidebar selection and triggers a transcript reload if
+    /// the sub-agent's turn count has advanced. Called once per render frame.
+    pub fn poll_metrics(&mut self) {
+        if let Some(ref mut rx) = self.metrics_rx
+            && rx.has_changed().unwrap_or(false)
+        {
+            let new_metrics = rx.borrow_and_update().clone();
+            // IC2: reset plan_view_active (subagents-override) when a new plan appears.
+            // Detect new plan by comparing graph_id; new plan should be shown immediately.
+            let new_graph_id = new_metrics
+                .orchestration_graph
+                .as_ref()
+                .map(|s| &s.graph_id);
+            let old_graph_id = self
+                .metrics
+                .orchestration_graph
+                .as_ref()
+                .map(|s| &s.graph_id);
+            if new_graph_id != old_graph_id && new_graph_id.is_some() {
+                self.sessions.current_mut().plan_view_active = false;
+            }
+            self.metrics = new_metrics;
+        }
+        // Clamp sidebar selection in case subagents count changed.
+        let count = self.metrics.sub_agents.len();
+        self.subagent_sidebar.clamp(count);
+        // Trigger transcript reload when turns count increased.
+        self.maybe_reload_transcript();
+    }
+
+    /// Evict oldest messages when the buffer exceeds `MAX_TUI_MESSAGES` (#2737).
+    ///
+    /// Shifts the render cache to match the drained messages, preserving cached renders
+    /// for the remaining entries and avoiding a full re-render stall (#2775).
+    pub(super) fn trim_messages(&mut self) {
+        self.sessions.current_mut().trim_messages();
+    }
+
+    /// Return a slice of all chat messages currently in the buffer.
+    ///
+    /// For the currently-displayed messages (which may be a sub-agent
+    /// transcript) use [`visible_messages`](Self::visible_messages) instead.
+    #[must_use]
+    pub fn messages(&self) -> &[ChatMessage] {
+        &self.sessions.current().messages
+    }
+
+    /// Return the current content of the text input field.
+    #[must_use]
+    pub fn input(&self) -> &str {
+        &self.sessions.current().input
+    }
+
+    /// Return the current input mode (normal vs. insert).
+    #[must_use]
+    pub fn input_mode(&self) -> InputMode {
+        self.sessions.current().input_mode
+    }
+
+    /// Return the cursor byte position within the input string.
+    #[must_use]
+    pub fn cursor_position(&self) -> usize {
+        self.sessions.current().cursor_position
+    }
+
+    /// Returns the composer height requested by the current draft, capped at three visible rows.
+    #[must_use]
+    pub(crate) fn desired_input_height(&self) -> u16 {
+        let content_lines = self.input_line_count().min(MAX_VISIBLE_INPUT_LINES);
+        content_lines.saturating_add(2)
+    }
+
+    /// Returns the number of logical lines in the current draft or indicator.
+    #[must_use]
+    pub(crate) fn input_line_count(&self) -> u16 {
+        if self.sessions.current().paste_state.is_some()
+            || (self.sessions.current().input.is_empty()
+                && matches!(self.sessions.current().input_mode, InputMode::Insert))
+        {
+            1
+        } else {
+            u16::try_from(self.sessions.current().input.matches('\n').count() + 1)
+                .unwrap_or(u16::MAX)
+        }
+    }
+
+    /// Return the number of lines the chat view is scrolled up from the bottom.
+    ///
+    /// `0` means the view is at the bottom (latest messages visible).
+    #[must_use]
+    pub fn scroll_offset(&self) -> usize {
+        self.sessions.current().scroll_offset
+    }
+
+    /// Scroll to bottom only if already at (or near) the bottom.
+    pub(super) fn auto_scroll(&mut self) {
+        if self.sessions.current().scroll_offset <= 1 {
+            self.sessions.current_mut().scroll_offset = 0;
+        }
+    }
+
+    /// Return `true` when tool-output blocks are expanded to full height.
+    #[must_use]
+    pub fn tool_expanded(&self) -> bool {
+        self.tool_expanded
+    }
+
+    /// Return the active paste indicator state, if any.
+    ///
+    /// `Some` when a multiline paste is in the input buffer and no edit
+    /// keypress has occurred since the paste. `None` otherwise.
+    #[must_use]
+    pub fn paste_state(&self) -> Option<&PasteState> {
+        self.sessions.current().paste_state.as_ref()
+    }
+
+    /// Return the current tool-output density level.
+    #[must_use]
+    pub fn tool_density(&self) -> ToolDensity {
+        self.tool_density
+    }
+
+    /// Return `true` when source-label badges are shown on assistant messages.
+    #[must_use]
+    pub fn show_source_labels(&self) -> bool {
+        self.show_source_labels
+    }
+
+    /// Toggle source-label visibility.
+    ///
+    /// Clears the render cache so all messages are re-rendered with the new
+    /// setting on the next frame.
+    pub fn set_show_source_labels(&mut self, v: bool) {
+        if self.show_source_labels != v {
+            self.show_source_labels = v;
+            self.sessions.current_mut().render_cache.clear();
+        }
+    }
+
+    /// Return `true` when the Cocoon TON balance should be shown in the status bar.
+    ///
+    /// Controlled by `[cocoon] show_balance` in config (default `true`). When `false`,
+    /// the balance is redacted to `*** TON` per spec §15.2.
+    #[must_use]
+    pub fn show_balance(&self) -> bool {
+        self.show_balance
+    }
+
+    /// Set whether the Cocoon TON balance is shown in the status bar.
+    pub fn set_show_balance(&mut self, v: bool) {
+        self.show_balance = v;
+    }
+
+    /// Replace the current hyperlink span list with `links`.
+    ///
+    /// Called by the render loop after each frame to store spans detected in
+    /// the terminal buffer so they can be emitted as OSC 8 sequences.
+    pub fn set_hyperlinks(&mut self, links: Vec<HyperlinkSpan>) {
+        self.hyperlinks = links;
+    }
+
+    /// Take ownership of the accumulated hyperlink spans, clearing the list.
+    ///
+    /// Called once per frame; the caller writes OSC 8 sequences to the terminal.
+    pub fn take_hyperlinks(&mut self) -> Vec<HyperlinkSpan> {
+        std::mem::take(&mut self.hyperlinks)
+    }
+
+    /// Return the current activity status label, if any.
+    ///
+    /// Displayed in the activity bar with a spinner when non-`None`
+    /// (e.g. `"Searching memory…"`, `"Executing tool: bash"`).
+    #[must_use]
+    pub fn status_label(&self) -> Option<&str> {
+        self.sessions.current().status_label.as_deref()
+    }
+
+    /// Return the number of messages queued or pending for the agent.
+    ///
+    /// Displayed in the input bar to indicate backpressure.
+    #[must_use]
+    pub fn queued_count(&self) -> usize {
+        self.queued_count.max(self.pending_count)
+    }
+
+    /// Return the projected context token count from the last assembly, or 0 if not yet known.
+    ///
+    /// The value is approximate (character-level heuristic) and is updated once per agent turn.
+    ///
+    /// # Examples
+    ///
+    /// ```rust
+    /// use tokio::sync::mpsc;
+    /// use zeph_tui::App;
+    ///
+    /// let (tx, _) = mpsc::channel(1);
+    /// let (_, rx) = mpsc::channel(1);
+    /// let app = App::new(tx, rx);
+    /// assert_eq!(app.context_token_estimate(), 0);
+    /// ```
+    #[must_use]
+    pub fn context_token_estimate(&self) -> usize {
+        self.context_token_estimate
+    }
+
+    /// Return `true` when the user is currently editing a queued message.
+    #[must_use]
+    pub fn editing_queued(&self) -> bool {
+        self.editing_queued
+    }
+
+    /// Return `true` when the agent is actively processing (streaming or running a tool).
+    ///
+    /// Used by the render loop to decide whether to show the activity spinner.
+    #[must_use]
+    pub fn is_agent_busy(&self) -> bool {
+        self.sessions.current().status_label.is_some()
+            || self
+                .sessions
+                .current()
+                .messages
+                .last()
+                .is_some_and(|m| m.streaming)
+    }
+
+    /// Return `true` when the last message is a streaming tool output.
+    #[must_use]
+    pub fn has_running_tool(&self) -> bool {
+        self.sessions
+            .current()
+            .messages
+            .last()
+            .is_some_and(|m| m.role == MessageRole::Tool && m.streaming)
+    }
+
+    /// Return a reference to the throbber animation state.
+    ///
+    /// Used by the status widget to render the spinner frame.
+    #[must_use]
+    pub fn throbber_state(&self) -> &throbber_widgets_tui::ThrobberState {
+        &self.throbber_state
+    }
+
+    /// Return a mutable reference to the throbber animation state.
+    ///
+    /// Called by the tick handler to advance the spinner frame each tick.
+    pub fn throbber_state_mut(&mut self) -> &mut throbber_widgets_tui::ThrobberState {
+        &mut self.throbber_state
+    }
+}
